+++ mobile/src/main/java/io/ashdavies/eternity/Reporting.java
@@ -1,0 +1,8 @@
+package io.ashdavies.eternity;
+
+import android.os.Bundle;
+
+public interface Reporting {
+
+  void log(String message, Bundle bundle);
+}