+++ mobile/src/main/java/io/ashdavies/eternity/firebase/FirebaseLogger.java
@@ -1,8 +1,5 @@
 package io.ashdavies.eternity.firebase;
 
-import android.content.Context;
-import android.os.Bundle;
-import com.google.firebase.analytics.FirebaseAnalytics;
 import com.google.firebase.crash.FirebaseCrash;
 import io.ashdavies.commons.util.StringUtils;
 import io.ashdavies.eternity.Logger;
@@ -10,26 +7,19 @@ import javax.inject.Inject;
 
 class FirebaseLogger implements Logger {
 
-  private final FirebaseAnalytics analytics;
-
   @Inject
-  FirebaseLogger(Context context) {
-    this(FirebaseAnalytics.getInstance(context));
-  }
-
-  private FirebaseLogger(FirebaseAnalytics analytics) {
-    this.analytics = analytics;
+  FirebaseLogger() {
   }
 
   @Override
-  public void log(String event, Bundle bundle) {
-    analytics.logEvent(event, bundle);
+  public void log(String message) {
+    FirebaseCrash.log(message);
   }
 
   @Override
   public void error(Throwable throwable, Object... args) {
     if (args != null) {
-      FirebaseCrash.log(StringUtils.toString(args));
+      log(StringUtils.toString(args));
     }
 
     FirebaseCrash.report(throwable);