+++ mobile/src/main/java/io/ashdavies/eternity/firebase/FirebaseModule.java
@@ -1,29 +1,28 @@
 package io.ashdavies.eternity.firebase;
 
-import android.content.Context;
-import android.content.SharedPreferences;
+import dagger.Binds;
 import dagger.Module;
 import dagger.Provides;
 import io.ashdavies.eternity.Config;
 import io.ashdavies.eternity.Logger;
+import io.ashdavies.eternity.Reporting;
 import io.ashdavies.eternity.inject.ApplicationScope;
 
 @Module
-public class FirebaseModule {
-
-  @Provides
-  Config.State state(SharedPreferences preferences) {
-    return new ConfigStateStorage(preferences);
-  }
+public abstract class FirebaseModule {
 
   @Provides
   @ApplicationScope
-  Config config(Config.State state) {
+  static Config config(Config.State state) {
     return FirebaseConfig.newInstance(state);
   }
 
-  @Provides
-  Logger logger(Context context) {
-    return new FirebaseLogger(context);
-  }
+  @Binds
+  abstract Config.State state(ConfigStateStorage storage);
+
+  @Binds
+  abstract Logger logger(FirebaseLogger logger);
+
+  @Binds
+  abstract Reporting reporting(FirebaseReporting reporting);
 }